List All Available Stability Measures

Lists all stability measures of package stabm and provides information about them.

listStabilityMeasures()

Value

data.frame

For each stability measure, its name, the information, whether it is corrected for chance by definition, the information, whether it is adjusted for similar features, its minimal value and its maximal value are displayed.

Note

The given minimal values might only be reachable in some scenarios, e.g. if the feature sets have a certain size. The measures which are not corrected for chance by definition can be corrected for chance with correction.for.chance. This however changes the minimal value. For the adjusted stability measures, the minimal value depends on the similarity structure.

Examples

listStabilityMeasures()
#>                           Name Corrected Adjusted Minimum Maximum
#> 1               stabilityDavis     FALSE    FALSE       0       1
#> 2                stabilityDice     FALSE    FALSE       0       1
#> 3             stabilityHamming     FALSE    FALSE       0       1
#> 4   stabilityIntersectionCount      TRUE     TRUE    <NA>       1
#> 5  stabilityIntersectionGreedy      TRUE     TRUE    <NA>       1
#> 6     stabilityIntersectionMBM      TRUE     TRUE    <NA>       1
#> 7    stabilityIntersectionMean      TRUE     TRUE    <NA>       1
#> 8             stabilityJaccard     FALSE    FALSE       0       1
#> 9               stabilityKappa      TRUE    FALSE      -1       1
#> 10         stabilityLustgarten      TRUE    FALSE      -1       1
#> 11           stabilityNogueira      TRUE    FALSE      -1       1
#> 12         stabilityNovovicova     FALSE    FALSE       0       1
#> 13             stabilityOchiai     FALSE    FALSE       0       1
#> 14                stabilityPhi      TRUE    FALSE      -1       1
#> 15           stabilitySechidis     FALSE     TRUE    <NA>      NA
#> 16              stabilitySomol      TRUE    FALSE       0       1
#> 17         stabilityUnadjusted      TRUE    FALSE      -1       1
#> 18               stabilityWald      TRUE    FALSE     1-p       1
#> 19                 stabilityYu      TRUE     TRUE    <NA>       1
#> 20           stabilityZucknick     FALSE     TRUE       0       1