Stability Measure Lustgarten

The stability of feature selection is defined as the robustness of the sets of selected features with respect to small variations in the data on which the feature selection is conducted. To quantify stability, several datasets from the same data generating process can be used. Alternatively, a single dataset can be split into parts by resampling. Either way, all datasets used for feature selection must contain exactly the same features. The feature selection method of interest is applied on all of the datasets and the sets of chosen features are recorded. The stability of the feature selection is assessed based on the sets of chosen features using stability measures.

stabilityLustgarten(features, p, impute.na = NULL)

Arguments

features
list (length >= 2)
Chosen features per dataset. Each element of the list contains the features for one dataset. The features must be given by their names (character) or indices (integerish).

p
numeric(1)
Total number of features in the datasets.

impute.na
numeric(1)
In some scenarios, the stability cannot be assessed based on all feature sets. E.g. if some of the feature sets are empty, the respective pairwise comparisons yield NA as result. With which value should these missing values be imputed? NULL means no imputation.

Value

numeric(1) Stability value.

Details

The stability measure is defined as (see Notation) $$\frac{2}{m (m - 1)} \sum_{i=1}^{m-1} \sum_{j = i+1}^m \frac{|V_i \cap V_j| - \frac{|V_i| \cdot |V_j|}{p}} {\min \{|V_i|, |V_j|\} - \max \{ 0, |V_i| + |V_j| - p \}}.$$

Notation

For the definition of all stability measures in this package, the following notation is used: Let \(V_1, \ldots, V_m\) denote the sets of chosen features for the \(m\) datasets, i.e. features has length \(m\) and \(V_i\) is a set which contains the \(i\)-th entry of features. Furthermore, let \(h_j\) denote the number of sets that contain feature \(X_j\) so that \(h_j\) is the absolute frequency with which feature \(X_j\) is chosen. Also, let \(q = \sum_{j=1}^p h_j = \sum_{i=1}^m |V_i|\) and \(V = \bigcup_{i=1}^m V_i\).

References

J. L. Lustgarten, V. Gopalakrishnan, and S. Visweswaran, "Measuring stability of feature selection in biomedical datasets", AMIA Annual Symposium proceedings/AMIA Symposium. AMIA Symposium, vol. 2009, pp. 406-410, 2009.

A. Bommert, J. Rahnenführer, and M. Lang, "A multi-criteria approach to find predictive and sparse models with stable feature selection for high-dimensional data", Computational and mathematical methods in medicine, 2017.

See also

Examples

feats = list(1:3, 1:4, 1:5)
stabilityLustgarten(features = feats, p = 10)

#> [1] 0.5333333